+++ CHANGELOG.md
@@ -57,6 +57,7 @@ And please only add new entries to the top of this list, right below the `# Unre
 - **Breaking:** On X11, device events are now ignored for unfocused windows by default, use `EventLoopWindowTarget::set_device_event_filter` to set the filter level.
 - Implemented `Default` on `EventLoop<()>`.
 - Implemented `Eq` for `Fullscreen`, `Theme`, and `UserAttentionType`.
+- On Windows, report the correct `DeviceId` for `WindowEvent::Touch`.
 
 # 0.26.1 (2022-01-05)